Remove unnecessary Gson Annotations from the domain model object "Breed"
