+++ app/src/main/java/com/cottacush/android/androidbaseprojectkt/sample/models/Breed.kt
@@ -1,36 +1,19 @@
 package com.cottacush.android.androidbaseprojectkt.sample.models
 import android.os.Parcelable
+import com.cottacush.android.androidbaseprojectkt.sample.Weight
 import com.google.gson.annotations.SerializedName
 import kotlinx.android.parcel.Parcelize
 
 @Parcelize
 data class Breed(
-    @SerializedName("id")
     val id: String,
-    @SerializedName("description")
     val description: String,
-    @SerializedName("life_span")
     val lifeSpan: String,
-    @SerializedName("name")
     val name: String,
-    @SerializedName("origin")
     val origin: String,
-    @SerializedName("social_needs")
     val socialNeeds: Int,
-    @SerializedName("stranger_friendly")
     val strangerFriendly: Int,
-    @SerializedName("temperament")
     val temperament: String,
-    @SerializedName("weight")
     val weight: Weight,
-    @SerializedName("wikipedia_url")
     val wikipediaUrl: String?
-) : Parcelable
-
-@Parcelize
-data class Weight(
-    @SerializedName("imperial")
-    val imperial: String,
-    @SerializedName("metric")
-    val metric: String
 ) : Parcelable
